requirements.txt
asgiref==3.7.2
crispy-tailwind==0.5.0
Django==4.2.7
django-crispy-forms==2.1
django-environ==0.11.2
django-js-asset==2.1.0
django-mptt==0.15.0
django-seed==0.3.1
django-treebeard==4.7
Faker==20.0.3
psycopg==3.1.12
psycopg-binary==3.1.12
python-dateutil==2.8.2
six==1.16.0
sqlparse==0.4.4
toposort==1.10
typing_extensions==4.8.0
